move function into utils file
